--- desktop.ini
@@ -1,3 +1,0 @@
-[LocalizedFileNames]
-glfw3.dll=@glfw3.dll,0
-glew32.dll=@glew32.dll,0
