--- trashcan.txt
@@ -1,37 +1,0 @@
- // Обновление пуль пришельцев
-            '''
-            for (size_t i = 0; i < GAME_MAX_ALIEN_BULLETS;) {
-                if (game.alien_bullets[i].is_active) {
-                    game.alien_bullets[i].y += game.alien_bullets[i].dir;
-                    if (game.alien_bullets[i].y < 0) { // Если пуля ушла за нижнюю границу
-                        game.alien_bullets[i].is_active = false;
-                    }
-                }
-                std::cout << "main game loop cycle update bullets of aliens\n";
-                if (!game.alien_bullets[i].is_active) {
-                    // Перемещаем последнюю активную пулю на место текущей неактивной
-                    if (i < game.num_alien_bullets - 1) { // Чтобы избежать копирования самой себя, если это последняя пуля
-                         // Найдем последнюю активную пулю и переместим ее
-                        size_t last_active_idx = (size_t)-1;
-                        for(size_t j = GAME_MAX_ALIEN_BULLETS -1; j >=0 ; --j){
-                            if(game.alien_bullets[j].is_active){
-                                last_active_idx = j;
-                                break;
-                            }
-                        }
-                        if(last_active_idx != (size_t)-1 && last_active_idx != i){
-                            game.alien_bullets[i] = game.alien_bullets[last_active_idx];
-                            game.alien_bullets[last_active_idx].is_active = false;
-                        } else {
-                            game.alien_bullets[i].is_active = false; // Просто деактивируем, если это последняя активная
-                        }
-
-                    } else { // Если это последняя пуля в активном списке
-                        game.alien_bullets[i].is_active = false;
-                    }
-                    --game.num_alien_bullets; // Уменьшаем счетчик
-                } else {
-                    ++i;
-                }
-            }
-            '''
